requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cachetools>=5.0.0
requests>=2.28.0
